improve code quality based on PR review
