+++ gson/src/main/java/com/google/gson/reflect/impl/UnsafeReflectionAccessor.java
@@ -27,7 +27,7 @@ import java.lang.reflect.Field;
  * NOTE: This implementation is designed for Java 9. Although it should work with earlier Java releases, it is better to
  * use {@link PreJava9ReflectionAccessor} for them.
  */
-public class UnsafeReflectionAccessor implements ReflectionAccessor {
+public final class UnsafeReflectionAccessor extends ReflectionAccessor {
 
   private static final Unsafe theUnsafe = getUnsafeInstance();
   private static final Field overrideField = getOverrideField();
